+++ lib/src/async/ime.dart
@@ -34,7 +34,8 @@ class IME {
   @override
   bool operator ==(other) =>
       other is IME &&
-      other._client == this._client && other._handler == this._handler;
+      other._client == this._client &&
+      other._handler == this._handler;
 
   @override
   String toString() => '$_handler.appManagemenet($_client)';