+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+- Add device lock related commands: `driver.device.*`
+- Add device system time command: `driver.device.getSystemTime()`
 
 ## [0.0.5] - 2019-06-22
 - Add context related commands: `driver.contexts.*`